+++ package.json
@@ -21,6 +21,7 @@
     "lucide-react": "^0.562.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-force-graph-2d": "^1.29.0",
     "react-quill": "^2.0.0",
     "react-router-dom": "^7.12.0",
     "zustand": "^4.4.7"
